# openapi.yaml
openapi: 3.0.0
info:
  title: Microblog
  description: Uma api para uma aplicação de microblog
  version: "0.1"
servers: 
  - url: http://localhost:8080/
paths:
...
  /api/users/{username}:
    get:
      summary: Retorna dados de um usuário
      description: Retorna as informações de um usuário
      parameters:
        - name: username
          in: path
          description: username do usuário
          required: true
          schema: 
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: um recurso usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
            description: Usuário não encontrado
  /api/users/update:
    put:
      summary: Atualiaza um Usuário
      description: Atualizado os dados do usuário autenticado
      requestBody:
        content: 
          application/json:
            schema: 
              $ref: '#/components/schemas/UserUpdate'
      security:
        - bearerAuth: []
      responses:
          '200':
            description: Usuário Alterado com sucesso        
          '400':
            description: Campos inválidos
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: jwt
  schemas:
...
    UserUpdate:
      type: object
      properties:
        username:
          type: string
          example: rodrigo
        email:
          type: string
          example: rodrigoaramburu@gmail.com
        password:
          type: string
          example: secredo!
    User:
      type: object
      properties:
        id:
          type: string
          example: d8228d83-dbb5-4086-8808-b1bd790eb1e8
        username:
          type: string
          example: rodrigo
        email:
          type: string
          example: rodrigoaramburu@gmail.com
        createAt: 
          type: string
          format: date-time
        photo:
          type: string
          example: http://localhost:8080/pictures/rodrigo
...