# src/main/java/br/dev/botecodigital/microblog/users/dto/UserUpdateDTO.java
package br.dev.botecodigital.microblog.users.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public class UserUpdateDTO {
	@NotBlank
	private String username;
	
	@NotBlank
	@Email
	private String email;
	
	@Size(min = 6)
	private String password;

	public void setUsername(String username) {
		this.username = username;
	}

	public void setEmail(String email) {
		this.email = email;
	}

	public void setPassword(String password) {
		this.password = password;
	}

	public String getUsername() {
		return username;
	}

	public String getEmail() {
		return email;
	}

	public String getPassword() {
		return password;
	}



}