# src/main/java/br/dev/botecodigital/microblog/users/model/User.java
package br.dev.botecodigital.microblog.users.model;

import java.time.LocalDateTime;
import java.util.UUID;

import jakarta.persistence.Column;
import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

@Entity
public class User {



	@Id
	@GeneratedValue(strategy = GenerationType.AUTO)
	private UUID id;
	
	@Column(nullable = false, unique = true)
	private @NotBlank String username;
	
	@Column(nullable = false, unique = true)
	private @NotBlank String email;
	
	@Column(nullable = false)
	@Size(min = 6)
	private @NotBlank String password;
	
	private LocalDateTime createAt;

	@Column(nullable = true)
	private String photo; 
	
	public User() {}
	
	public User(@NotBlank String username, @NotBlank String email, @NotBlank String password) {
		this.username = username;
		this.email = email;
		this.password = password;
		
		this.createAt = LocalDateTime.now();
	}


	public UUID getId() {
		return id;
	}


	public void setId(UUID id) {
		this.id = id;
	}


	public String getUsername() {
		return username;
	}


	public void setUsername(String username) {
		this.username = username;
	}


	public String getEmail() {
		return email;
	}


	public void setEmail(String email) {
		this.email = email;
	}


	public String getPassword() {
		return password;
	}


	public void setPassword(String password) {
		this.password = password;
	}


	public LocalDateTime getCreateAt() {
		return createAt;
	}


	public void setCreateAt(LocalDateTime createAt) {
		this.createAt = createAt;
	}


	public String getPhoto() {
		return this.photo;
	}
	public void setPhoto(String photo) {
		this.photo = photo;
	}
	
	
}